+++ api/src/main/java/org/pmiops/workbench/vwb/wsm/WsmConfig.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.vwb.wsm.wsm;
+package org.pmiops.workbench.vwb.wsm;
 
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
