+++ api/src/test/java/org/pmiops/workbench/api/CohortReviewControllerTest.java
@@ -145,7 +145,7 @@ import org.pmiops.workbench.workspaces.WorkspaceServiceFactory;
 import org.pmiops.workbench.workspaces.WorkspaceServiceImpl;
 import org.pmiops.workbench.workspaces.resources.UserRecentResourceService;
 import org.pmiops.workbench.workspaces.resources.WorkspaceResourcesService;
-import org.pmiops.workbench.vwb.wsm.wsm.WsmClient;
+import org.pmiops.workbench.vwb.wsm.WsmClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
